feat(topology): introduce filters, topological spaces and uniform spaces
This also adds smaller theories of finite sets, cartesian product of sets,
complete boolean algebras.

The goal is to constructs the completion of a uniform space. And provide a
mechanism to extends functions into complete spaces from dense subsets to the
closure of this dense subset. This is the basis to introduce real numbers.